factor out personal info box
